pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "beadloom"
version = "0.1.0"
description = "Context Oracle + Doc Sync Engine for AI-assisted development"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Beadloom Contributors" },
]
keywords = ["ai", "context", "documentation", "mcp", "knowledge-graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Documentation",
    "Typing :: Typed",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "tree-sitter>=0.23",
    "tree-sitter-python>=0.23",
    "mcp>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "types-PyYAML>=6.0",
]

[project.scripts]
beadloom = "beadloom.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/beadloom"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

[tool.ruff]
target-version = "py310"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
    "PTH",  # flake8-use-pathlib
    "S",    # flake8-bandit
    "T20",  # flake8-print
]
ignore = [
    "S101",  # assert used in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
packages = ["beadloom"]
mypy_path = "src"

[tool.coverage.run]
source = ["beadloom"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 80
